requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy.orm
pydantic
pwdlib[argon2]
hashlib
python-dotenv   
PyJWT 
python-multipart
pydantic[email]
fastapi-mail